+++ src/main/java/at/hannibal2/skyhanni/SkyHanniMod.kt
@@ -398,14 +398,13 @@ class SkyHanniMod {
         loadModule(DungeonCleanEnd())
         loadModule(DungeonBossMessages())
         loadModule(DungeonBossHideDamageSplash())
-        loadModule(TrophyFishManager())
+        loadModule(TrophyFishManager)
         loadModule(TrophyFishFillet())
         loadModule(TrophyFishMessages())
         loadModule(BazaarBestSellMethod())
         loadModule(BazaarOpenPriceWebsite())
         loadModule(AnvilCombineHelper())
         loadModule(SeaCreatureMessageShortener())
-        //        registerEvent(new GriffinBurrowFinder());
         loadModule(AshfangFreezeCooldown())
         loadModule(AshfangNextResetCooldown())
         loadModule(SummoningSoulsName())
