+++ src/main/java/at/hannibal2/skyhanni/config/commands/Commands.kt
@@ -211,7 +211,7 @@ object Commands {
         ) { CaptureFarmingGear.reverseCarrotFortune() }
         registerCommand(
             "shrepostatus",
-        "Shows the status of all the mods constants"
+            "Shows the status of all the mods constants"
         ) { SkyHanniMod.repo.displayRepoStatus(false) }
     }
